+++ src/main/java/atwoz/atwoz/admin/application/AdminAuthService.java
@@ -11,8 +11,8 @@ import atwoz.atwoz.admin.domain.admin.Admin;
 import atwoz.atwoz.admin.domain.admin.Password;
 import atwoz.atwoz.admin.domain.repository.AdminRepository;
 import atwoz.atwoz.admin.domain.service.PasswordHasher;
-import atwoz.atwoz.common.auth.context.Role;
-import atwoz.atwoz.common.auth.jwt.JwtProvider;
+import atwoz.atwoz.auth.context.Role;
+import atwoz.atwoz.auth.jwt.JwtProvider;
 import atwoz.atwoz.common.domain.vo.Email;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
