+++ src/main/java/atwoz/atwoz/auth/config/SecurityConfig.java
@@ -1,4 +1,4 @@
-package atwoz.atwoz.common.auth.config;
+package atwoz.atwoz.auth.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
