+++ src/main/java/atwoz/atwoz/auth/context/AuthContext.java
@@ -1,4 +1,4 @@
-package atwoz.atwoz.common.auth.context;
+package atwoz.atwoz.auth.context;
 
 import lombok.AccessLevel;
 import lombok.Getter;
