+++ src/main/java/atwoz/atwoz/auth/filter/extractor/RefreshTokenExtractor.java
@@ -1,4 +1,4 @@
-package atwoz.atwoz.common.auth.filter.extractor;
+package atwoz.atwoz.auth.filter.extractor;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
