+++ src/main/java/atwoz/atwoz/auth/jwt/JwtParser.java
@@ -1,6 +1,6 @@
-package atwoz.atwoz.common.auth.jwt;
+package atwoz.atwoz.auth.jwt;
 
-import atwoz.atwoz.common.auth.context.Role;
+import atwoz.atwoz.auth.context.Role;
 import io.jsonwebtoken.*;
 import io.jsonwebtoken.security.Keys;
 import org.springframework.beans.factory.annotation.Value;
