--- src/main/java/atwoz/atwoz/common/auth/jwt/JwtRepository.java
@@ -1,25 +1,0 @@
-package atwoz.atwoz.common.auth.jwt;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.stereotype.Repository;
-
-import java.util.concurrent.TimeUnit;
-
-@Repository
-@RequiredArgsConstructor
-public class JwtRepository {
-    private final RedisTemplate<String, String> redisTemplate;
-
-    public void save(String token) {
-        redisTemplate.opsForValue().set(token, "", 1, TimeUnit.DAYS);
-    }
-
-    public void delete(String token) {
-        redisTemplate.delete(token);
-    }
-
-    public boolean isExists(String token) {
-        return redisTemplate.hasKey(token);
-    }
-}
