+++ src/main/java/atwoz/atwoz/member/presentation/MemberController.java
@@ -1,7 +1,7 @@
 package atwoz.atwoz.member.presentation;
 
-import atwoz.atwoz.common.auth.context.AuthContext;
-import atwoz.atwoz.common.auth.context.AuthPrincipal;
+import atwoz.atwoz.auth.context.AuthContext;
+import atwoz.atwoz.auth.context.AuthPrincipal;
 import atwoz.atwoz.common.presentation.BaseResponse;
 import atwoz.atwoz.common.presentation.StatusType;
 import atwoz.atwoz.member.application.MemberService;
