+++ src/test/java/atwoz/atwoz/admin/application/AdminAuthServiceTest.java
@@ -11,8 +11,8 @@ import atwoz.atwoz.admin.domain.admin.Admin;
 import atwoz.atwoz.admin.domain.admin.InvalidPasswordException;
 import atwoz.atwoz.admin.domain.repository.AdminRepository;
 import atwoz.atwoz.admin.domain.service.PasswordHasher;
-import atwoz.atwoz.common.auth.context.Role;
-import atwoz.atwoz.common.auth.jwt.JwtProvider;
+import atwoz.atwoz.auth.context.Role;
+import atwoz.atwoz.auth.jwt.JwtProvider;
 import atwoz.atwoz.common.domain.vo.Email;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
