+++ src/test/java/atwoz/atwoz/common/auth/filter/TokenFilterTest.java
@@ -1,11 +1,12 @@
 package atwoz.atwoz.common.auth.filter;
 
-import atwoz.atwoz.common.auth.context.AuthContext;
-import atwoz.atwoz.common.auth.context.Role;
-import atwoz.atwoz.common.auth.filter.response.ResponseHandler;
-import atwoz.atwoz.common.auth.jwt.JwtParser;
-import atwoz.atwoz.common.auth.jwt.JwtProvider;
-import atwoz.atwoz.common.auth.jwt.JwtRepository;
+import atwoz.atwoz.auth.context.AuthContext;
+import atwoz.atwoz.auth.context.Role;
+import atwoz.atwoz.auth.filter.TokenFilter;
+import atwoz.atwoz.auth.filter.response.ResponseHandler;
+import atwoz.atwoz.auth.jwt.JwtParser;
+import atwoz.atwoz.auth.jwt.JwtProvider;
+import atwoz.atwoz.auth.jwt.JwtRepository;
 import atwoz.atwoz.common.presentation.StatusType;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
